refactor: move local storage functions
