+++ src/context/context.tsx
@@ -14,6 +14,7 @@ import { getHeaders, getOAuthHeaders } from '@/lib/auth';
 import { removeSearchParams, replaceSearchParams } from '@/lib/route';
 import { PlayerView } from '@/types/state';
 import { createContext } from 'react';
+import { getLsFollowedGames, setLsFollowedGames } from '@/lib/local-storage';
 
 const LS_ACCESS_TOKEN = 'ACCESS_TOKEN';
 const VALIDATE_INTERVAL = 60 * 60 * 1000;
@@ -238,56 +239,6 @@ export function AppProvider({ children }: { children: ReactNode }) {
     setFollowedGames(games);
   }, []);
 
-  const SEPARATOR = '___';
-  function getLsFollowedGames(): FollowedGame[] {
-    // try get from storage
-    let gamesStr = localStorage.getItem(LS_FOLLOWED_GAMES);
-    if (!gamesStr) return [];
-
-    // parse string
-    let followedGames: FollowedGame[] = [];
-    let rows = gamesStr.split('\n');
-    for (let r of rows) {
-      let game = r.split(SEPARATOR);
-
-      if (game.length != 2) continue;
-      let gameId = Number.parseInt(game[0]);
-      if (Number.isNaN(gameId)) continue;
-
-      let gameTitle = game[1];
-      followedGames.push({ game_id: gameId, game_title: gameTitle });
-    }
-    console.log('testestest', followedGames);
-
-    return followedGames;
-  }
-
-  function setLsFollowedGames(games: FollowedGame[] | undefined) {
-    // test
-    games = [
-      { game_id: 1229, game_title: 'Super Mario World' },
-      { game_id: 505705, game_title: 'Noita' },
-    ];
-
-    if (!games) {
-      localStorage.removeItem(LS_FOLLOWED_GAMES);
-      return;
-    }
-
-    // to string
-    let followedGamesStr = '';
-    for (let i = 0; i < games.length; ++i) {
-      if (i > 0) {
-        followedGamesStr += '\n';
-      }
-      let g = games[i];
-      let gameStr = g.game_id + SEPARATOR + g.game_title;
-      followedGamesStr += gameStr;
-    }
-
-    // save to local storage
-    localStorage.setItem(LS_FOLLOWED_GAMES, followedGamesStr);
-  }
   return (
     <AppContext.Provider
       value={{
